requirements.txt
django
django-rest-framework
django-import-export